+++ crates/system-env/src/pm_vendor.rs
@@ -8,11 +8,62 @@ macro_rules! string_vec {
     }};
 }
 
+#[derive(Clone, Debug)]
+pub struct ListParser {
+    regex: regex::Regex,
+}
+
+impl ListParser {
+    pub fn new(pattern: &str) -> Self {
+        let pattern = pattern
+            .replace("<package>", "(?<package>[a-zA-Z0-9-_]+)")
+            .replace(
+                "<version>",
+                "(?<version>(\\d+)(\\.\\d+)?(\\.\\d+)?(-[a-z0-9-_+.]+)?)",
+            )
+            .replace("<distro>", "(?<distro>[a-zA-Z0-9-_,]+)")
+            .replace("<tag>", "(?<tag>@[0-9.]+)?")
+            .replace("<arch>", "(x86_64|amd64|aarch64|arm64)");
+
+        ListParser {
+            regex: regex::Regex::new(&pattern).unwrap(),
+        }
+    }
+
+    pub fn parse(&self, output: &str) -> HashMap<String, Option<String>> {
+        let mut packages = HashMap::default();
+
+        for line in output.lines() {
+            let line = line.trim();
+
+            if line.starts_with('#') || line.starts_with("//") {
+                continue;
+            }
+
+            if let Some(caps) = self.regex.captures(line) {
+                let Some(name) = caps.name("package") else {
+                    continue;
+                };
+
+                packages.insert(
+                    name.as_str().to_string(),
+                    caps.name("version").map(|cap| cap.as_str().to_string()),
+                );
+            }
+        }
+
+        packages
+    }
+}
+
 /// The types of commands that are currently supported by package managers.
 #[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
 pub enum CommandType {
     /// Installs a system dependency.
     InstallPackage,
+
+    /// Return all installed dependencies.
+    ListPackages,
 
     /// Updates the registry index.
     UpdateIndex,
@@ -47,6 +98,9 @@ pub struct PackageManagerConfig {
     /// Mapping of command types to CLI arguments.
     pub commands: HashMap<CommandType, Vec<String>>,
 
+    /// Parser for extracting information from installed lists.
+    pub list_parser: ListParser,
+
     /// How interactive/prompt arguments are handled.
     pub prompt_arg: PromptArgument,
 
@@ -61,14 +115,21 @@ pub(crate) fn apk() -> PackageManagerConfig {
     PackageManagerConfig {
         commands: HashMap::from_iter([
             (CommandType::InstallPackage, string_vec!["apk", "add", "$"]),
+            // https://www.cyberciti.biz/faq/alpine-linux-apk-list-files-in-package/
+            (
+                CommandType::ListPackages,
+                string_vec!["apk", "list", "--installed"],
+            ),
             (CommandType::UpdateIndex, string_vec!["apk", "update"]),
         ]),
+        list_parser: ListParser::new("^<package>-<version>"),
         prompt_arg: PromptArgument::Interactive("-i".into()),
         prompt_for: vec![CommandType::InstallPackage, CommandType::UpdateIndex],
         version_arg: VersionArgument::Inline("=".into()),
     }
 }
 
+// https://manpages.ubuntu.com/manpages/xenial/man8/apt.8.html
 pub(crate) fn apt() -> PackageManagerConfig {
     PackageManagerConfig {
         commands: HashMap::from_iter([
@@ -76,8 +137,14 @@ pub(crate) fn apt() -> PackageManagerConfig {
                 CommandType::InstallPackage,
                 string_vec!["apt", "install", "--install-recommends", "$"],
             ),
+            // https://www.cyberciti.biz/faq/ubuntu-lts-debian-linux-apt-command-examples/#12
+            (
+                CommandType::ListPackages,
+                string_vec!["apt", "list", "--installed"],
+            ),
             (CommandType::UpdateIndex, string_vec!["apt", "update"]),
         ]),
+        list_parser: ListParser::new("^<package>/<distro> <version>"),
         prompt_arg: PromptArgument::Skip("-y".into()),
         prompt_for: vec![CommandType::InstallPackage, CommandType::UpdateIndex],
         version_arg: VersionArgument::Inline("=".into()),
@@ -91,8 +158,13 @@ pub(crate) fn brew() -> PackageManagerConfig {
                 CommandType::InstallPackage,
                 string_vec!["brew", "install", "$"],
             ),
+            (
+                CommandType::ListPackages,
+                string_vec!["brew", "list", "--formula", "--versions"],
+            ),
             (CommandType::UpdateIndex, string_vec!["brew", "update"]),
         ]),
+        list_parser: ListParser::new("^<package><tag> <version>"),
         prompt_arg: PromptArgument::None, // Interactive("-i".into()),
         prompt_for: vec![CommandType::InstallPackage],
         version_arg: VersionArgument::Inline("@".into()),
@@ -101,10 +173,15 @@ pub(crate) fn brew() -> PackageManagerConfig {
 
 pub(crate) fn choco() -> PackageManagerConfig {
     PackageManagerConfig {
-        commands: HashMap::from_iter([(
-            CommandType::InstallPackage,
-            string_vec!["choco", "install", "$"],
-        )]),
+        commands: HashMap::from_iter([
+            (
+                CommandType::InstallPackage,
+                string_vec!["choco", "install", "$"],
+            ),
+            // https://docs.chocolatey.org/en-us/choco/commands/list/
+            (CommandType::ListPackages, string_vec!["choco", "list"]),
+        ]),
+        list_parser: ListParser::new("^<package> <version>"),
         prompt_arg: PromptArgument::Skip("-y".into()),
         prompt_for: vec![CommandType::InstallPackage],
         version_arg: VersionArgument::Separate("--version".into()),
@@ -118,8 +195,14 @@ pub(crate) fn dnf() -> PackageManagerConfig {
                 CommandType::InstallPackage,
                 string_vec!["dnf", "install", "$"],
             ),
+            // https://tylersguides.com/guides/listing-installed-packages-with-dnf/
+            (
+                CommandType::ListPackages,
+                string_vec!["dnf", "list", "installed"],
+            ),
             (CommandType::UpdateIndex, string_vec!["dnf", "check-update"]),
         ]),
+        list_parser: ListParser::new("^<package>.<arch>\\s+<version>"),
         prompt_arg: PromptArgument::Skip("-y".into()),
         prompt_for: vec![CommandType::InstallPackage, CommandType::UpdateIndex],
         version_arg: VersionArgument::Inline("-".into()),
@@ -133,14 +216,18 @@ pub(crate) fn pacman() -> PackageManagerConfig {
                 CommandType::InstallPackage,
                 string_vec!["pacman", "-S", "$"],
             ),
+            // https://www.atlantic.net/dedicated-server-hosting/list-installed-packages-with-pacman-on-arch-linux/
+            (CommandType::ListPackages, string_vec!["pacman", "-Q"]),
             (CommandType::UpdateIndex, string_vec!["pacman", "-Syy"]),
         ]),
+        list_parser: ListParser::new("^<package> <version>"),
         prompt_arg: PromptArgument::Skip("--noconfirm".into()),
         prompt_for: vec![CommandType::InstallPackage],
         version_arg: VersionArgument::Inline(">=".into()),
     }
 }
 
+// https://man.freebsd.org/cgi/man.cgi?query=pkg&sektion=8&manpath=freebsd-release-ports
 pub(crate) fn pkg() -> PackageManagerConfig {
     PackageManagerConfig {
         commands: HashMap::from_iter([
@@ -148,8 +235,14 @@ pub(crate) fn pkg() -> PackageManagerConfig {
                 CommandType::InstallPackage,
                 string_vec!["pkg", "install", "$"],
             ),
+            // https://www.zenarmor.com/docs/freebsd-tutorials/how-to-install-packages-with-pkg-on-freebsd
+            (
+                CommandType::ListPackages,
+                string_vec!["pkg", "info", "--all"],
+            ),
             (CommandType::UpdateIndex, string_vec!["pkg", "update"]),
         ]),
+        list_parser: ListParser::new("^<package>-<version>"),
         prompt_arg: PromptArgument::Skip("-y".into()),
         prompt_for: vec![CommandType::InstallPackage],
         version_arg: VersionArgument::None,
@@ -165,6 +258,7 @@ pub(crate) fn pkg() -> PackageManagerConfig {
 //     }
 // }
 
+// https://pkgin.net/
 pub(crate) fn pkgin() -> PackageManagerConfig {
     PackageManagerConfig {
         commands: HashMap::from_iter([
@@ -172,8 +266,11 @@ pub(crate) fn pkgin() -> PackageManagerConfig {
                 CommandType::InstallPackage,
                 string_vec!["pkgin", "install", "$"],
             ),
+            // https://www.unitedbsd.com/d/571-pkgin-statslists-return-seems-false/29
+            (CommandType::ListPackages, string_vec!["pkgin", "list"]),
             (CommandType::UpdateIndex, string_vec!["pkgin", "update"]),
         ]),
+        list_parser: ListParser::new("^<package>-<version>"),
         prompt_arg: PromptArgument::Skip("-y".into()),
         prompt_for: vec![CommandType::InstallPackage, CommandType::UpdateIndex],
         version_arg: VersionArgument::Inline("-".into()),
@@ -187,8 +284,11 @@ pub(crate) fn scoop() -> PackageManagerConfig {
                 CommandType::InstallPackage,
                 string_vec!["scoop", "install", "$"],
             ),
+            // https://rcjach.github.io/blog/scoop/#scoop-list
+            (CommandType::ListPackages, string_vec!["scoop", "list"]),
             (CommandType::UpdateIndex, string_vec!["scoop", "update"]),
         ]),
+        list_parser: ListParser::new("^<package> <version>"),
         prompt_arg: PromptArgument::None,
         prompt_for: vec![],
         version_arg: VersionArgument::Inline("@".into()),
@@ -202,8 +302,14 @@ pub(crate) fn yum() -> PackageManagerConfig {
                 CommandType::InstallPackage,
                 string_vec!["yum", "install", "$"],
             ),
+            // https://phoenixnap.com/kb/how-to-list-installed-packages-on-centos
+            (
+                CommandType::ListPackages,
+                string_vec!["yum", "list", "installed"],
+            ),
             (CommandType::UpdateIndex, string_vec!["yum", "check-update"]),
         ]),
+        list_parser: ListParser::new("^<package>.<arch>\\s+<version>"),
         prompt_arg: PromptArgument::Skip("-y".into()),
         prompt_for: vec![CommandType::InstallPackage],
         version_arg: VersionArgument::Inline("-".into()),