+++ crates/system-env/src/system.rs
@@ -54,31 +54,22 @@ impl System {
         dep_config: &DependencyConfig,
         interactive: bool,
     ) -> Result<Option<Vec<String>>, Error> {
-        let Some(pm) = dep_config.manager.or(self.manager) else {
-            return Err(Error::RequiredPackageManager);
-        };
-
-        let pm_config = pm.get_config();
-        let mut args = vec![];
-
-        let Some(base_args) = pm_config.commands.get(&CommandType::InstallPackage) else {
+        let Some(base_args) = self.extract_command(CommandType::InstallPackage, interactive)?
+        else {
             return Ok(None);
         };
 
+        let mut args = vec![];
+        let pm = self.manager.as_ref().unwrap();
+        let pm_config = pm.get_config();
+
         for arg in base_args {
             if arg == "$" {
-                args.extend(self.extract_package_args(dep_config, &pm_config, &pm)?);
+                args.extend(self.extract_package_args(dep_config, &pm_config, pm)?);
             } else {
-                args.push(arg.to_owned());
+                args.push(arg);
             }
         }
-
-        self.append_interactive(
-            CommandType::InstallPackage,
-            &pm_config,
-            &mut args,
-            interactive,
-        );
 
         Ok(Some(args))
     }
@@ -94,35 +85,35 @@ impl System {
         dep_configs: &[DependencyConfig],
         interactive: bool,
     ) -> Result<Option<Vec<String>>, Error> {
-        let Some(pm) = self.manager else {
-            return Err(Error::RequiredPackageManager);
-        };
-
-        let pm_config = pm.get_config();
-        let mut args = vec![];
-
-        let Some(base_args) = pm_config.commands.get(&CommandType::InstallPackage) else {
+        let Some(base_args) = self.extract_command(CommandType::InstallPackage, interactive)?
+        else {
             return Ok(None);
         };
+
+        let mut args = vec![];
+        let pm = self.manager.as_ref().unwrap();
+        let pm_config = pm.get_config();
 
         for arg in base_args {
             if arg == "$" {
                 for dep_config in dep_configs {
-                    args.extend(self.extract_package_args(dep_config, &pm_config, &pm)?);
+                    args.extend(self.extract_package_args(dep_config, &pm_config, pm)?);
                 }
             } else {
-                args.push(arg.to_owned());
+                args.push(arg);
             }
         }
 
-        self.append_interactive(
-            CommandType::InstallPackage,
-            &pm_config,
-            &mut args,
-            interactive,
-        );
-
         Ok(Some(args))
+    }
+
+    /// Return the command and arguments to "list installed packages"
+    /// for the current package manager.
+    pub fn get_list_packages_command(
+        &self,
+        interactive: bool,
+    ) -> Result<Option<Vec<String>>, Error> {
+        self.extract_command(CommandType::ListPackages, interactive)
     }
 
     /// Return the command and arguments to "update the registry index"
@@ -131,21 +122,7 @@ impl System {
         &self,
         interactive: bool,
     ) -> Result<Option<Vec<String>>, Error> {
-        let Some(pm) = self.manager else {
-            return Err(Error::RequiredPackageManager);
-        };
-
-        let pm_config = pm.get_config();
-
-        if let Some(args) = pm_config.commands.get(&CommandType::UpdateIndex) {
-            let mut args = args.to_owned();
-
-            self.append_interactive(CommandType::UpdateIndex, &pm_config, &mut args, interactive);
-
-            return Ok(Some(args));
-        }
-
-        Ok(None)
+        self.extract_command(CommandType::UpdateIndex, interactive)
     }
 
     /// Resolve and reduce the dependencies to a list that's applicable
@@ -200,6 +177,28 @@ impl System {
                 }
             };
         }
+    }
+
+    fn extract_command(
+        &self,
+        command: CommandType,
+        interactive: bool,
+    ) -> Result<Option<Vec<String>>, Error> {
+        let Some(pm) = self.manager else {
+            return Err(Error::RequiredPackageManager);
+        };
+
+        let pm_config = pm.get_config();
+
+        if let Some(args) = pm_config.commands.get(&command) {
+            let mut args = args.to_owned();
+
+            self.append_interactive(command, &pm_config, &mut args, interactive);
+
+            return Ok(Some(args));
+        }
+
+        Ok(None)
     }
 
     fn extract_package_args(