+++ crates/system-env/tests/pm_test.rs
@@ -5,7 +5,7 @@ fn one_dep() -> DependencyConfig {
 }
 
 fn many_dep() -> DependencyConfig {
-    SystemDependency::names(["foo", "bar", "baz"]).to_config()
+    SystemDependency::names(["bar", "baz", "foo"]).to_config()
 }
 
 mod pm {
@@ -36,13 +36,13 @@ mod pm {
                 pm.get_install_package_command(&many_cfg, false)
                     .unwrap()
                     .unwrap(),
-                vec!["apk", "add", "foo", "bar", "baz"]
+                vec!["apk", "add", "bar", "baz", "foo"]
             );
             assert_eq!(
                 pm.get_install_package_command(&many_cfg, true)
                     .unwrap()
                     .unwrap(),
-                vec!["apk", "add", "foo", "bar", "baz", "-i"]
+                vec!["apk", "add", "bar", "baz", "foo", "-i"]
             );
         }
 
@@ -57,6 +57,20 @@ mod pm {
                     .unwrap()
                     .unwrap(),
                 vec!["apk", "add", "foo=1.2.3"]
+            );
+        }
+
+        #[test]
+        fn list_packages() {
+            let pm = System::with_manager(SystemPackageManager::Apk);
+
+            assert_eq!(
+                pm.get_list_packages_command(false).unwrap().unwrap(),
+                vec!["apk", "list", "--installed"]
+            );
+            assert_eq!(
+                pm.get_list_packages_command(true).unwrap().unwrap(),
+                vec!["apk", "list", "--installed"]
             );
         }
 
@@ -104,9 +118,9 @@ mod pm {
                     "apt",
                     "install",
                     "--install-recommends",
-                    "foo",
                     "bar",
                     "baz",
+                    "foo",
                     "-y"
                 ]
             );
@@ -118,9 +132,9 @@ mod pm {
                     "apt",
                     "install",
                     "--install-recommends",
-                    "foo",
                     "bar",
-                    "baz"
+                    "baz",
+                    "foo",
                 ]
             );
         }
@@ -136,6 +150,20 @@ mod pm {
                     .unwrap()
                     .unwrap(),
                 vec!["apt", "install", "--install-recommends", "foo=1.2.3", "-y"]
+            );
+        }
+
+        #[test]
+        fn list_packages() {
+            let pm = System::with_manager(SystemPackageManager::Apt);
+
+            assert_eq!(
+                pm.get_list_packages_command(false).unwrap().unwrap(),
+                vec!["apt", "list", "--installed"]
+            );
+            assert_eq!(
+                pm.get_list_packages_command(true).unwrap().unwrap(),
+                vec!["apt", "list", "--installed"]
             );
         }
 
@@ -179,13 +207,13 @@ mod pm {
                 pm.get_install_package_command(&many_cfg, false)
                     .unwrap()
                     .unwrap(),
-                vec!["brew", "install", "foo", "bar", "baz"]
+                vec!["brew", "install", "bar", "baz", "foo"]
             );
             assert_eq!(
                 pm.get_install_package_command(&many_cfg, true)
                     .unwrap()
                     .unwrap(),
-                vec!["brew", "install", "foo", "bar", "baz"], // , "-i"]
+                vec!["brew", "install", "bar", "baz", "foo"], // , "-i"]
             );
         }
 
@@ -200,6 +228,20 @@ mod pm {
                     .unwrap()
                     .unwrap(),
                 vec!["brew", "install", "foo@1.2.3"]
+            );
+        }
+
+        #[test]
+        fn list_packages() {
+            let pm = System::with_manager(SystemPackageManager::Brew);
+
+            assert_eq!(
+                pm.get_list_packages_command(false).unwrap().unwrap(),
+                vec!["brew", "list", "--formula", "--versions"]
+            );
+            assert_eq!(
+                pm.get_list_packages_command(true).unwrap().unwrap(),
+                vec!["brew", "list", "--formula", "--versions"]
             );
         }
 
@@ -243,13 +285,13 @@ mod pm {
                 pm.get_install_package_command(&many_cfg, false)
                     .unwrap()
                     .unwrap(),
-                vec!["choco", "install", "foo", "bar", "baz", "-y"]
+                vec!["choco", "install", "bar", "baz", "foo", "-y"]
             );
             assert_eq!(
                 pm.get_install_package_command(&many_cfg, true)
                     .unwrap()
                     .unwrap(),
-                vec!["choco", "install", "foo", "bar", "baz"]
+                vec!["choco", "install", "bar", "baz", "foo"]
             );
         }
 
@@ -264,6 +306,20 @@ mod pm {
                     .unwrap()
                     .unwrap(),
                 vec!["choco", "install", "foo", "--version", "1.2.3", "-y"]
+            );
+        }
+
+        #[test]
+        fn list_packages() {
+            let pm = System::with_manager(SystemPackageManager::Choco);
+
+            assert_eq!(
+                pm.get_list_packages_command(false).unwrap().unwrap(),
+                vec!["choco", "list"]
+            );
+            assert_eq!(
+                pm.get_list_packages_command(true).unwrap().unwrap(),
+                vec!["choco", "list"]
             );
         }
 
@@ -300,13 +356,13 @@ mod pm {
                 pm.get_install_package_command(&many_cfg, false)
                     .unwrap()
                     .unwrap(),
-                vec!["dnf", "install", "foo", "bar", "baz", "-y"]
+                vec!["dnf", "install", "bar", "baz", "foo", "-y"]
             );
             assert_eq!(
                 pm.get_install_package_command(&many_cfg, true)
                     .unwrap()
                     .unwrap(),
-                vec!["dnf", "install", "foo", "bar", "baz"]
+                vec!["dnf", "install", "bar", "baz", "foo"]
             );
         }
 
@@ -321,6 +377,20 @@ mod pm {
                     .unwrap()
                     .unwrap(),
                 vec!["dnf", "install", "foo-1.2.3", "-y"]
+            );
+        }
+
+        #[test]
+        fn list_packages() {
+            let pm = System::with_manager(SystemPackageManager::Dnf);
+
+            assert_eq!(
+                pm.get_list_packages_command(false).unwrap().unwrap(),
+                vec!["dnf", "list", "installed"]
+            );
+            assert_eq!(
+                pm.get_list_packages_command(true).unwrap().unwrap(),
+                vec!["dnf", "list", "installed"]
             );
         }
 
@@ -364,13 +434,13 @@ mod pm {
                 pm.get_install_package_command(&many_cfg, false)
                     .unwrap()
                     .unwrap(),
-                vec!["pacman", "-S", "foo", "bar", "baz", "--noconfirm"]
+                vec!["pacman", "-S", "bar", "baz", "foo", "--noconfirm"]
             );
             assert_eq!(
                 pm.get_install_package_command(&many_cfg, true)
                     .unwrap()
                     .unwrap(),
-                vec!["pacman", "-S", "foo", "bar", "baz"]
+                vec!["pacman", "-S", "bar", "baz", "foo"]
             );
         }
 
@@ -385,6 +455,20 @@ mod pm {
                     .unwrap()
                     .unwrap(),
                 vec!["pacman", "-S", "foo>=1.2.3", "--noconfirm"]
+            );
+        }
+
+        #[test]
+        fn list_packages() {
+            let pm = System::with_manager(SystemPackageManager::Pacman);
+
+            assert_eq!(
+                pm.get_list_packages_command(false).unwrap().unwrap(),
+                vec!["pacman", "-Q"]
+            );
+            assert_eq!(
+                pm.get_list_packages_command(true).unwrap().unwrap(),
+                vec!["pacman", "-Q"]
             );
         }
 
@@ -428,13 +512,13 @@ mod pm {
                 pm.get_install_package_command(&many_cfg, false)
                     .unwrap()
                     .unwrap(),
-                vec!["pkg", "install", "foo", "bar", "baz", "-y"]
+                vec!["pkg", "install", "bar", "baz", "foo", "-y"]
             );
             assert_eq!(
                 pm.get_install_package_command(&many_cfg, true)
                     .unwrap()
                     .unwrap(),
-                vec!["pkg", "install", "foo", "bar", "baz"]
+                vec!["pkg", "install", "bar", "baz", "foo"]
             );
         }
 
@@ -449,6 +533,20 @@ mod pm {
                     .unwrap()
                     .unwrap(),
                 vec!["pkg", "install", "foo", "-y"]
+            );
+        }
+
+        #[test]
+        fn list_packages() {
+            let pm = System::with_manager(SystemPackageManager::Pkg);
+
+            assert_eq!(
+                pm.get_list_packages_command(false).unwrap().unwrap(),
+                vec!["pkg", "info", "--all"]
+            );
+            assert_eq!(
+                pm.get_list_packages_command(true).unwrap().unwrap(),
+                vec!["pkg", "info", "--all"]
             );
         }
 
@@ -492,13 +590,13 @@ mod pm {
                 pm.get_install_package_command(&many_cfg, false)
                     .unwrap()
                     .unwrap(),
-                vec!["pkgin", "install", "foo", "bar", "baz", "-y"]
+                vec!["pkgin", "install", "bar", "baz", "foo", "-y"]
             );
             assert_eq!(
                 pm.get_install_package_command(&many_cfg, true)
                     .unwrap()
                     .unwrap(),
-                vec!["pkgin", "install", "foo", "bar", "baz"]
+                vec!["pkgin", "install", "bar", "baz", "foo"]
             );
         }
 
@@ -513,6 +611,20 @@ mod pm {
                     .unwrap()
                     .unwrap(),
                 vec!["pkgin", "install", "foo-1.2.3", "-y"]
+            );
+        }
+
+        #[test]
+        fn list_packages() {
+            let pm = System::with_manager(SystemPackageManager::Pkgin);
+
+            assert_eq!(
+                pm.get_list_packages_command(false).unwrap().unwrap(),
+                vec!["pkgin", "list"]
+            );
+            assert_eq!(
+                pm.get_list_packages_command(true).unwrap().unwrap(),
+                vec!["pkgin", "list"]
             );
         }
 
@@ -556,13 +668,13 @@ mod pm {
                 pm.get_install_package_command(&many_cfg, false)
                     .unwrap()
                     .unwrap(),
-                vec!["scoop", "install", "foo", "bar", "baz"]
+                vec!["scoop", "install", "bar", "baz", "foo"]
             );
             assert_eq!(
                 pm.get_install_package_command(&many_cfg, true)
                     .unwrap()
                     .unwrap(),
-                vec!["scoop", "install", "foo", "bar", "baz"]
+                vec!["scoop", "install", "bar", "baz", "foo"]
             );
         }
 
@@ -577,6 +689,20 @@ mod pm {
                     .unwrap()
                     .unwrap(),
                 vec!["scoop", "install", "foo@1.2.3"]
+            );
+        }
+
+        #[test]
+        fn list_packages() {
+            let pm = System::with_manager(SystemPackageManager::Scoop);
+
+            assert_eq!(
+                pm.get_list_packages_command(false).unwrap().unwrap(),
+                vec!["scoop", "list"]
+            );
+            assert_eq!(
+                pm.get_list_packages_command(true).unwrap().unwrap(),
+                vec!["scoop", "list"]
             );
         }
 
@@ -620,13 +746,13 @@ mod pm {
                 pm.get_install_package_command(&many_cfg, false)
                     .unwrap()
                     .unwrap(),
-                vec!["yum", "install", "foo", "bar", "baz", "-y"]
+                vec!["yum", "install", "bar", "baz", "foo", "-y"]
             );
             assert_eq!(
                 pm.get_install_package_command(&many_cfg, true)
                     .unwrap()
                     .unwrap(),
-                vec!["yum", "install", "foo", "bar", "baz"]
+                vec!["yum", "install", "bar", "baz", "foo"]
             );
         }
 
@@ -641,6 +767,20 @@ mod pm {
                     .unwrap()
                     .unwrap(),
                 vec!["yum", "install", "foo-1.2.3", "-y"]
+            );
+        }
+
+        #[test]
+        fn list_packages() {
+            let pm = System::with_manager(SystemPackageManager::Yum);
+
+            assert_eq!(
+                pm.get_list_packages_command(false).unwrap().unwrap(),
+                vec!["yum", "list", "installed"]
+            );
+            assert_eq!(
+                pm.get_list_packages_command(true).unwrap().unwrap(),
+                vec!["yum", "list", "installed"]
             );
         }